feat(api/): updated api structure
