+++ api/src/controllers/Admin/Category/createCategory.js
@@ -1,6 +1,7 @@
-const  Category  = require('../models/Category');
+const Category = require('../../models/Category');
 
-const create = (req, res, next) => {
+module.exports = {
+  createCategory: (req, res, next) => {
     const { name } = req.body;
     if (!name) {
         return res.status(404).send('Debes ingresar Nombre');
@@ -10,17 +11,5 @@ const create = (req, res, next) => {
         if (err) return res.status(400).json(err);
         res.json(newCategory);
     });
-
+  }
 }
-
-const get = (req, res, next) => {
-    Category.find({}).then(Categories => {
-        return res.json(Categories);
-    })
-}
-
-module.exports = {
-    create,
-    get
-}
-
