+++ api/src/controllers/Admin/Category/getCategories.js
@@ -1,0 +1,9 @@
+const Category = require('../../models/Category');
+
+module.exports = {
+  getCategories: (req, res) => {
+    Category.find({}).then(Categories => {
+        return res.json(Categories);
+    });
+  }
+}
