+++ api/src/controllers/General/getAllProducts.js
@@ -1,12 +1,12 @@
 const Product = require("../models/Product");
 
-async function searchP (req, res, next) {
+module.exports = {
+  getAllProducts: async (req, res, next) => {
     try {
-        let product = await Product.find();
+        const product = await Product.find();
         return res.json(product);
       } catch (error) {
         console.log(error);
       }
-    }
-
-module.exports = searchP
+  }
+}
